指摘された部分の修正

+++ app/src/main/java/com/ginga/naviai/admin/dto/StatsResponse.java
@@ -1,14 +1,14 @@
 package com.ginga.naviai.admin.dto;
 
 public class StatsResponse {
-    private int pending;
-    private int published;
-    private int declined;
+    private long pending;
+    private long published;
+    private long declined;
     public StatsResponse() {}
-    public int getPending() { return pending; }
-    public void setPending(int pending) { this.pending = pending; }
-    public int getPublished() { return published; }
-    public void setPublished(int published) { this.published = published; }
-    public int getDeclined() { return declined; }
-    public void setDeclined(int declined) { this.declined = declined; }
+    public long getPending() { return pending; }
+    public void setPending(long pending) { this.pending = pending; }
+    public long getPublished() { return published; }
+    public void setPublished(long published) { this.published = published; }
+    public long getDeclined() { return declined; }
+    public void setDeclined(long declined) { this.declined = declined; }
 }
